chore: extract rsbuild / rspack config modify into rsbuild plugin
